+++ CHANGES.md
@@ -15,6 +15,7 @@
 ### Improvements
 
 -   Capture all errors logged by gdal when opening a file fails (#495).
+-   Improve support for datetime columns (#486).
 -   Add support to read and write ".gpkg.zip" (GDAL >= 3.7), ".shp.zip", and ".shz"
     files (#527).
 -   Compatibility with the string dtype in the upcoming pandas 3.0 release (#493).